TwitterAPI Factory Objects created.
Attacks to API should be done from a more complex and abstract layer, so we create this factory to abstract more complex api operation, specially connections and settings with the api.
In the future concurrents use of API should be treated in this layer.

Objects instantiate by this factory will encapsulate the api queries calls.